Moved code around.
