+++ tools/parse/src/main/scala/org/allenai/nlpstack/parse/poly/fsm/SculptureTrainingVectorSource.scala
@@ -28,7 +28,7 @@ case class SculptureTrainingVectorSource(
       taggedSentence <- sculptureSource.sculptureIterator
       vector <- generateVectors(taggedSentence)
     } yield {
-      println(vector)
+      //println(vector)
       vector
     }
   }
